requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Brotli==1.0.9
click==8.1.3
dj-database-url==1.2.0
Django==4.1.6
django-bootstrap-v5==1.0.11
django-filter==22.1
django-htmx==1.13.0
django-tables2==2.5.2
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
gunicorn==20.1.0
mypy-extensions==1.0.0
openpyxl==3.1.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
python-decouple==3.7
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
tomli==2.0.1
typing_extensions==4.4.0
whitenoise==6.3.0
xlrd==2.0.1
xlwt==1.3.0
